refactor(Layout): move Menu component under Layout folder
